+++ coreutils/length.c
@@ -9,6 +9,6 @@ extern int length_main(int argc, char **argv)
 	if (argc != 2 || **(argv + 1) == '-') {
 		usage("length string\n");
 	}
-	printf("%d\n", strlen(argv[1]));
+	printf("%lu\n", (long)strlen(argv[1]));
 	return (TRUE);
 }
